requirements.txt
flask
flask-login
flask-wtf
werkzeug
langchain
langchain-core
langchain-community
langchain-huggingface
docx2txt
pypdf
langchain_chroma
sentence-transformers
llama-cpp-python
mysql-connector-python
huggingface-hub
flask-cors
requests
pydub
elevenlabs
openai-whisper
python-dotenv
gunicorn